Combined the rest of the class to the class array
This is so that the element will only have one class attribute using the
class binding.  Much cleaner.  This way you also have the ability to
specify in which part the class can be applied. In this example, the
top and part class gets applied first then the sale-border. With the
array syntax, I can put the sale-border in between or in front.